refactor so can run from terminal properly
