+++ naturalperm/isomorphism.py
@@ -1,5 +1,4 @@
 from typing import Protocol
-from abc import abstractmethod
 
 class Isomorphism(Protocol):
